Extract PostgresScriptGenerator from PostgresDatabase (#3)
Moves all model-to-SQL generation out of PostgresDatabase into a standalone,
connection-free PostgresScriptGenerator. PostgresDatabase.PublishAsync now
delegates to it. Adds fast Docker-free unit tests that assert on generated SQL
for table, index, and unique-index-with-method/direction cases.

Also fixes a builder divergence surfaced by the new tests: the parser stored
the varchar Length property as long while the DB builder and generator used
int, so scripting a parser-built model threw InvalidCastException. Parser now
stores int, matching the other two.

+++ Squill.Provider.Postgres/ParserWorkspaceModelBuilder.cs
@@ -165,8 +165,10 @@ public class ParserWorkspaceModelBuilder : IDatabaseModelBuilder
                         {
                             throw new InvalidOperationException("Unexpected length modifier for varchar or character type");
                         }
-                        
-                        typeSpec.Properties.Add(new Property(PostgresPropertyNames.Length, length));
+
+                        // Store as int to match the DB-extraction builder and the script
+                        // generator, which both use int for the Length property.
+                        typeSpec.Properties.Add(new Property(PostgresPropertyNames.Length, (int)length));
                     }
                     else
                     {
